pembaruan data

+++ database/migrations/2024_05_09_073428_create_pakans_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->foreignId('doc_id')->constrained('docs')->onDelete('cascade')->onUpdate('cascade');
             $table->date('tanggal');
             $table->enum('jenis_pakan', ['Galaxy 00', 'Galaxy 01', 'B 151 C / Strarfeed', '510', '511 Alfa']); // Adjust enum values as needed
-            $table->integer('jumlah_pakan');
+            $table->float('jumlah_pakan');
             $table->integer('hrg_pakan_satuan');
             $table->integer('total_harga');
             $table->timestamps(); // Adds created_at and updated_at columns
